docs: restore search plugin
In f41f6925007350030316 a plugin list was added to the mkdocs config
to add the "social" plugin. This removed the default "search" plugin
(I assume by accident).

See https://www.mkdocs.org/user-guide/configuration/#plugins for the
default plugins.

The plugin is quite handy for searching audit rules, etc., so fix it
by adding "search" back to the enabled plugins.