refactor: add StepBodyCommon
This will allow me to remove the `uses()` and `with()`
trait methods, which encode unnecessary runtime error
states.

More fully, this implies that the `Uses` model needs
to be pushed into `github_actions_models` and made into
a hard parsing invariant (i.e. we should fail to parse
if a `uses:` clause is invalid).