chore(ci): Collect lint configuration in workspace root
Previously, a particular allow-by-default Clippy lint was denied in CI.
Unfortunately, developers wouldn't see it locally because it is
allow-by-default. This was also inconsistent with the dev docs. To fix
this, migrate the configuration of Clippy lints to the
`[workspace.lints]` section of the workspace manifest and configure all
crates to use that table. `[workspace.lints]` was stabilized in [Rust
1.74].

[Rust 1.74]: https://blog.rust-lang.org/2023/11/16/Rust-1.74.0/#lint-configuration-through-cargo