fix: treat empty config path as unset
We had a workflow start failing when we updated Zizmor to a version
which contained `ZIZMOR_CONFIG` support. The reason is that we have some
logic (in a CI job) which supplies a config file dynamically using
`--config`, and this logic uses the same environment variable name.

The problem was that the logic was setting `ZIZMOR_CONFIG` to an empty
string when we didn't have a config file to supply. This was previously
fine, but since 1e123cfbe5ad7534cf7c099fd21e74293880876b Zizmor itself
now looks at this variable. An empty string is treated as valid, so we
try to load a config file from the empty path, which fails.

```console
$ ZIZMOR_CONFIG= zizmor .
fatal: no audit was performed
error: failed to load config: No such file or directory (os error 2)
 = note: check your configuration file for errors
 = note: see: https://docs.zizmor.sh/configuration/
```

This can be a bit obscure to track down and fix, so what we do here is
treat the empty string the same as if the variable (or flag) was unset,
so we don't try to load a config file in this scenario.