refactor: AuditState::gh_client
Move the client as a property on `AuditState` to check token validity
(if given) early on. This triggers the check even if audit rules making
use of GH API are not used.

Additionally, instead of creating a new `Client` for audit rules needing
it, we create it once and clone it.